Change log for this commit
- Changed the event identifier parsing slightly by searching from right side. 
This will help in making arbitrary event strings later on with a common event
group as the right most substring.

- Added check in case the '-' could not be found at all. Had not thought of this 
case before.

- Added all the alphabets for the keyboard events. XD

- Added a TrivialEvent struct. All event structures must inherit from this. Will
help in some interesting typecasting among event objects. Read the code for more
explanation on how these structs are structured (no pun intended).

- Removing keyBoardEventMap. It wasn't being used anywhere. Not sure why it was 
present in the first place.

- Added up and down mouse button events along with a lot of code handling event
identifier parsing. Some of this will come in handy for making events for 
collisions. Read the code comments for more information on this. Probably the 
biggest part of all.

git-svn-id: http://projects.hashstash.in/svn/trivial@8 659a0b78-bc15-4dc9-aeea-a5f01ff3c5d0